+++ src/AnomalyDetectors/IsolationForest.php
@@ -182,7 +182,7 @@ class IsolationForest implements Estimator, Learner, Ranking, Persistable
         if (isset($this->contamination)) {
             $scores = array_map([self::class, 'isolationScore'], $dataset->samples());
 
-            $threshold = Stats::percentile($scores, 100. - (100. * $this->contamination));
+            $threshold = Stats::percentile($scores, 100. * (1. - $this->contamination));
         }
 
         $this->threshold = $threshold ?? self::DEFAULT_THRESHOLD;