+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - Added global transpose array helper function
     - Renamed label key to classes in Multiclass Breakdown report
     - Changed order of Gradient Boost and AdaBoost hyper-parameters
+    - Added NaN Euclidean distance kernel
 
 - 0.0.15-beta
     - Added Gaussian MLE anomaly detector